+++ src/client/IClient.ts
@@ -1,4 +1,4 @@
-import {CommentData, FullProjectData, SectionData} from '../types';
+import {CommentData, FullProjectData, SectionData} from '../types/music_sniper_types';
 
 export interface IClient {
     fetchFullDataForProject: (projectId: string) => Promise<FullProjectData | Error>;
