+++ src/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
+
 import './index.css';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
 import {LocalStorageStore} from './store/LocalStorageStore';
 import {LocalStorageClient} from './client/LocalStorageClient';
 
@@ -27,8 +27,3 @@ window.addEventListener('load', async () => {
         </React.StrictMode>
     );
 });
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
